% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum.R
\name{agepregprev_lf}
\alias{agepregprev_lf}
\title{Age-specific prevalence among pregnant women}
\usage{
agepregprev_lf(
  mod,
  fp,
  aidx = 3:9 * 5 - fp$ss$AGE_START + 1L,
  yidx = 1:fp$proj_years,
  agspan = 5,
  expand = FALSE
)
}
\arguments{
\item{mod}{todo}

\item{fp}{Leapfrog fixed parameters}

\item{expand}{whether to expand aidx, yidx, sidx, and agspan}
}
\description{
Age-specific prevalence among pregnant women
}
