% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum.R
\name{ageprev_lf}
\alias{ageprev_lf}
\title{Prevalence by arbitrary age groups}
\usage{
ageprev_lf(
  mod,
  a_idx = NULL,
  s_idx = NULL,
  y_idx = NULL,
  ag_span = 5,
  expand = FALSE,
  version = "C"
)
}
\arguments{
\item{sidx}{sex (1 = Male, 2 = Female, 0 = Both)
Notes: Assumes that AGE_START is 15 and single year of age.}
}
\description{
Prevalence by arbitrary age groups
}
