% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{ancsite_pred_df}
\alias{ancsite_pred_df}
\title{Prepare design matrix indices for ANC prevalence predictions}
\usage{
ancsite_pred_df(ancsite_df, fp)
}
\arguments{
\item{ancsite_df}{data.frame of site-level ANC design for predictions}

\item{fp}{fixed parameter input list}
}
\description{
Prepare design matrix indices for ANC prevalence predictions
}
\examples{
pjnz <- system.file("extdata/testpjnz", "Botswana2017.PJNZ", package="eppasm.lf")
bw <- prepare_spec_fit(pjnz, proj.end=2021.5)


bw_u_ancsite <- attr(bw$Urban, "eppd")$ancsitedat
fp <- attr(bw$Urban, "specfp")

ancsite_pred_df(bw_u_ancsite, fp)

}
