% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{ancsite_pred_df_lf}
\alias{ancsite_pred_df_lf}
\title{Prepare design matrix indices for ANC prevalence predictions}
\usage{
ancsite_pred_df_lf(ancsite_df, params)
}
\arguments{
\item{ancsite_df}{data.frame of site-level ANC design for predictions}

\item{params}{Leapfrog parameter inputs}
}
\description{
Prepare design matrix indices for ANC prevalence predictions
}
