% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-outputs.R
\name{create_aggr_outputs}
\alias{create_aggr_outputs}
\title{Crate aggr outputs}
\usage{
create_aggr_outputs(fitlist, fitnat, both = TRUE)
}
\arguments{
\item{fitlist}{list of model fits to aggregate.}

\item{fitnat}{single fit to use as the base fit with aggregation.}

\item{both}{flag indicating whether to return outputs for both aggregated fit and individual fits (default TRUE).}
}
\description{
Crate aggr outputs
}
