% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eppasm-outputs.R
\name{hivpop_singleage}
\alias{hivpop_singleage}
\title{Convert aggregate HIV population to single year}
\usage{
hivpop_singleage(mod, ss)
}
\arguments{
\item{mod}{EPP-ASM model output of class `spec`}

\item{ss}{Model state space inputs}
}
\description{
EPP-ASM tracks the CD4 distribution and ART duration
of the HIV population by coarse age groups `15-16, 17-19,
20-24, ..., 45-49, 50+` for computational efficiency.
}
\details{
`hivpop_singleage` converts the coarse age group CD4
distribution and ART coverage to single year of age counts
assuming uniform proportions in each category within coarse
age groups.
}
\examples{
pjnz <- system.file("extdata/testpjnz", "Botswana2018.PJNZ", package="eppasm.lf")
fp <- prepare_directincid(pjnz)
mod <- simmod(fp)
hivp1 <- hivpop_singleage(mod, fp$ss)

}
