% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imis.R
\name{imis}
\alias{imis}
\title{Incremental Mixture Importance Sampling (IMIS)}
\usage{
imis(
  B0,
  B,
  B_re,
  number_k,
  opt_k = NULL,
  fp,
  likdat,
  prior = eppasm.lf:::prior,
  likelihood = eppasm.lf:::likelihood,
  sample_prior = eppasm.lf:::sample.prior,
  dsamp = eppasm.lf:::dsamp,
  save_all = FALSE
)
}
\arguments{
\item{B0}{number of initial samples to draw}

\item{B}{number of samples at each IMIS iteration}

\item{B_re}{number of resamples}

\item{number_k}{maximum number of iterations}

\item{opt_k}{vector of iterations at which to use optimization step to identify new mixture component}

\item{fp}{fixed model parameters}

\item{likdat}{likeihood data}

\item{prior}{function to calculate prior density for matrix of parameter inputs}

\item{likelihood}{function to calculate likelihood for matrix of parameter inputs}

\item{sample_prior}{function to draw an initial sample of parameter inputs}

\item{dsamp}{function to calculate density for initial sampling distribution (may be equal to prior)}

\item{save_all}{logical whether to save all sampled parameters}
}
\value{
list with items resample, stat, and center
}
\description{
Implements IMIS algorithm with optional optimization step (Raftery and Bao 2010).
}
