% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{ll_hhsage_binom_lf}
\alias{ll_hhsage_binom_lf}
\title{Log likelihood for age-specific household survey prevalence using binomial
approximation}
\usage{
ll_hhsage_binom_lf(mod, fp, dat, pointwise = FALSE)
}
\description{
Log likelihood for age-specific household survey prevalence using binomial
approximation
}
