% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-prep.R
\name{prep_fp_fitmod}
\alias{prep_fp_fitmod}
\title{Prepare fp object for fitting and prepare likelihood data}
\usage{
prep_fp_fitmod(obj, ..., epp = FALSE)
}
\arguments{
\item{obj}{specfp or eppfp object}

\item{...}{Additional parameters to override on `obj`}

\item{epp}{If TRUE prep EPP model fit}
}
\value{
List containing `fp` object ready for fitting and likelihood
  data `likdat`
}
\description{
Prepare fp object for fitting and prepare likelihood data
}
