% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-prep.R
\name{prep_fp_fitmod_lf}
\alias{prep_fp_fitmod_lf}
\title{Prepare leapfrog inputs object for fitting and prepare likelihood data}
\usage{
prep_fp_fitmod_lf(params, eppd, epp_t0, ...)
}
\arguments{
\item{params}{Leapfrog input data}

\item{eppd}{EPP input data}

\item{...}{Additional parameters to set/override on `params`}
}
\value{
List containing `fp` object ready for fitting and likelihood
  data `likdat`
}
\description{
Prepare leapfrog inputs object for fitting and prepare likelihood data
}
