% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{prepare_hhsageprev_likdat_lf}
\alias{prepare_hhsageprev_likdat_lf}
\title{Prepare age-specific HH survey prevalence likelihood data}
\usage{
prepare_hhsageprev_likdat_lf(hhsage, params)
}
\description{
Prepare age-specific HH survey prevalence likelihood data
}
