% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{prepare_likdat_lf}
\alias{prepare_likdat_lf}
\title{Prepare likelihood data from leapfrog data}
\usage{
prepare_likdat_lf(params, eppd)
}
\arguments{
\item{params}{Leapfrog input data}

\item{eppd}{EPP input data}
}
\value{
List containing likelihood data
}
\description{
Prepare likelihood data from leapfrog data
}
