% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt-models.R
\name{prepare_rhybrid_lf}
\alias{prepare_rhybrid_lf}
\title{Setup the r-hybrid model}
\usage{
prepare_rhybrid_lf(
  params,
  ts_epidemic_start = 1975.5,
  rw_start = params$rw_start,
  rw_trans = params$rw_trans,
  rw_dk = params$rw_dk
)
}
\arguments{
\item{rw_start}{time when random walk starts}

\item{rw_trans}{number of years to transition from logistic differences to RW differences. If NULL, defaults to 5 steps}

\item{fp}{Leapfrog parameters object}

\item{tsEpidemicStart}{time step at which epidemic is seeded}
}
\description{
Setup the r-hybrid model
}
