% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-spectrum-files.R
\name{read_epp_perc_urban}
\alias{read_epp_perc_urban}
\title{Read percentage urban input from EPP XML file}
\usage{
read_epp_perc_urban(pjnz)
}
\arguments{
\item{pjnz}{file path to Spectrum PJNZ file.}
}
\description{
Read percentage urban input from EPP XML file
}
