% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-spectrum-files.R
\name{read_epp_t0}
\alias{read_epp_t0}
\title{Read epidemic start year from EPP XML file}
\usage{
read_epp_t0(pjnz)
}
\arguments{
\item{pjnz}{file path to Spectrum PJNZ file.}
}
\value{
vector of epidemic start year for each EPP subregion with region names
}
\description{
Read epidemic start year from EPP XML file
}
