% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-spectrum-files.R
\name{read_subp_file}
\alias{read_subp_file}
\title{Read subpopulation size input file}
\usage{
read_subp_file(filepath)
}
\arguments{
\item{filepath}{file path to .subp file}
}
\description{
Read subpopulation size input file
}
